+++ Runtime/EpisodeStatus.cs
@@ -1,0 +1,9 @@
+namespace Unity.AI.MLAgents
+{
+    public enum EpisodeStatus
+    {
+        DEFAULT,
+        TERMINATED,
+        INTERRUPTED
+    }
+}